set underlying interface for feedtails to limit pjs scope
