+++ apps/account-worker/src/transaction_publisher/publisher.service.ts
@@ -7,7 +7,8 @@ import { SubmittableExtrinsic } from '@polkadot/api-base/types';
 import { IMethod, ISubmittableResult, Signer, SignerResult } from '@polkadot/types/types';
 import { MILLISECONDS_PER_SECOND } from 'time-constants';
 import { SchedulerRegistry } from '@nestjs/schedule';
-import { BlockchainService, ICapacityInfo } from '#blockchain/blockchain.service';
+import { BlockchainService } from '#blockchain/blockchain.service';
+import { ICapacityInfo } from '#blockchain/types';
 import { AccountQueues as QueueConstants } from '#types/constants/queue.constants';
 import { BaseConsumer } from '#consumer';
 import { TransactionData } from '#types/dtos/account';
