+++ libs/blockchain/src/capacity-checker.service.ts
@@ -3,7 +3,7 @@ import { Injectable, Logger } from '@nestjs/common';
 import { InjectRedis } from '@songkeys/nestjs-redis';
 import { Redis } from 'ioredis';
 import { EventEmitter2 } from '@nestjs/event-emitter';
-import { ICapacityInfo } from './blockchain.service';
+import { ICapacityInfo } from './types';
 import { ConfigService } from '@nestjs/config';
 import { IBlockchainConfig } from './blockchain.config';
 import { BlockchainRpcQueryService } from './blockchain-rpc-query.service';
